+++ src/app/index.js
@@ -1,7 +1,10 @@
 import { createApp } from './app';
 import Vitessce from './Vitessce';
+import { encodeConfAsURLParams, decodeURLParamsToConf } from './export-utils';
 
 export {
   createApp,
   Vitessce,
+  encodeConfAsURLParams,
+  decodeURLParamsToConf,
 };